requirements.txt
aiogram==3.6.0
Babel==2.13.1
fastapi==0.111.0
python-decouple==3.8
uvicorn==0.29.0
